Simplify useThingOwnership
